fix: CI lint エラーを解消 (max-params, max-lines)
- CriticAuditor のコンストラクタを options オブジェクトパターンに変更 (max-params)
- reactive-layer.spec.ts をヘルパー抽出 + ファイル分割 (max-lines)
- stream-helpers.spec.ts を2ファイルに分割 (max-lines)
