+++ apps/discord/src/bootstrap.ts
@@ -306,7 +306,7 @@ export async function setupMemoryRecording(
 						storage = new MemoryStorage(resolveMemoryDbPath(dataDir, namespace));
 						storageCache.set(userId, storage);
 					}
-					const auditor = new CriticAuditor(llm, storage, driftCalculator, characterDefinition);
+					const auditor = new CriticAuditor({ llm, storage, driftCalculator, characterDefinition });
 					return auditor.audit(userId);
 				},
 			};
